restmusic/requirements.txt
Django==2.0
django-environ==0.4.5
django-webpack-loader==0.6.0
djangorestframework==3.9.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pytz==2018.7
